Refactoring for better organization of files
