+++ Ethenyl/SoftServe/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:SoftServe"
-             StartupUri="MainWindow.xaml">
+             StartupUri="Views/MainWindow.xaml">
     <Application.Resources>
         <Style x:Key="MediaButtons" TargetType="Button">
             <Setter Property="FontFamily" Value="Segoe MDL2 Assets, Segoe UI Symbol"/>
