+++ Ethenyl/SoftServe/Views/SettingsWindow.xaml.cs
@@ -1,20 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
+﻿using System.Windows;
 using Newtonsoft.Json;
 using SoftServe.ViewModels;
 
-namespace SoftServe
+namespace SoftServe.Views
 {
     /// <summary>
     /// Interaction logic for SettingsWindow.xaml
