+++ Ethenyl/SoftServe/packages.config
@@ -2,4 +2,9 @@
 <packages>
   <package id="Newtonsoft.Json" version="8.0.2" targetFramework="net452" />
   <package id="SpotifyAPI-NET" version="2.8.0" targetFramework="net452" />
+  <package id="sqlite-net-pcl" version="1.1.2" targetFramework="net452" />
+  <package id="SQLitePCL.bundle_green" version="0.9.2" targetFramework="net452" />
+  <package id="SQLitePCL.native.sqlite3.v110_xp" version="0.9.2" targetFramework="net452" />
+  <package id="SQLitePCL.plugin.sqlite3.net45" version="0.9.2" targetFramework="net452" />
+  <package id="SQLitePCL.raw" version="0.9.2" targetFramework="net452" />
 </packages>
